fix: console recorder imports are exploding bundle size
