+++ packages/record/package.json
@@ -56,7 +56,8 @@
   },
   "dependencies": {
     "@rrweb/types": "^2.0.0-alpha.18",
-    "rrweb": "^2.0.0-alpha.18"
+    "rrweb": "^2.0.0-alpha.18",
+    "@rrweb/utils": "^2.0.0-alpha.18"
   },
   "browserslist": [
     "supports es6-class"
