+++ packages/record/tsconfig.json
@@ -11,6 +11,9 @@
     },
     {
       "path": "../rrweb"
+    },
+    {
+      "path": "../utils"
     }
   ]
 }
