+++ packages/rrweb/src/utils.ts
@@ -127,49 +127,6 @@ export function hookSetter<T>(
   return () => hookSetter(target, key, original || {}, true);
 }
 
-// copy from https://github.com/getsentry/sentry-javascript/blob/b2109071975af8bf0316d3b5b38f519bdaf5dc15/packages/utils/src/object.ts
-export function patch(
-  source: { [key: string]: any },
-  name: string,
-  replacement: (...args: unknown[]) => unknown,
-): () => void {
-  try {
-    if (!(name in source)) {
-      return () => {
-        //
-      };
-    }
-
-    const original = source[name] as () => unknown;
-    const wrapped = replacement(original);
-
-    // Make sure it's a function first, as we need to attach an empty prototype for `defineProperties` to work
-    // otherwise it'll throw "TypeError: Object.defineProperties called on non-object"
-    if (typeof wrapped === 'function') {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      wrapped.prototype = wrapped.prototype || {};
-      Object.defineProperties(wrapped, {
-        __rrweb_original__: {
-          enumerable: false,
-          value: original,
-        },
-      });
-    }
-
-    source[name] = wrapped;
-
-    return () => {
-      source[name] = original;
-    };
-  } catch {
-    return () => {
-      //
-    };
-    // This can throw if multiple fill happens on a global object like XMLHttpRequest
-    // Fixes https://github.com/getsentry/sentry-javascript/issues/2043
-  }
-}
-
 // guard against old third party libraries which redefine Date.now
 let nowTimestamp = Date.now;
